Windows

+++ src/par2cmdline.h
@@ -52,8 +52,8 @@ typedef unsigned char    u8;
 typedef signed char      i8;
 typedef unsigned short   u16;
 typedef signed short     i16;
-typedef unsigned long    u32;
-typedef signed long      i32;
+typedef unsigned int     u32;
+typedef signed int       i32;
 typedef unsigned __int64 u64;
 typedef signed __int64   i64;
